==> R/01_load.qmd <==
---
title: "Load"
editor: visual
message: false
format:
  html:
    embed-resources: true
---

## Load

```{r message = FALSE}
#Load the necessary libraries for data handling and manipulation


library(tidyverse)
library(readr)
source("../R/99_proj_func.R") #Custom functions created by us
```

```{r}
# Generates data folder and _raw sub folder
generate_dir() #call function to generate the directory
```

```{r message=FALSE}
# Here we define the URLs for downloading human and mouse data files

human_file <- "https://raw.githubusercontent.com/s232996/project/main/gCSI_GRdata_v1.3.tsv/data/human.csv"

human_suppl_file <- "https://raw.githubusercontent.com/s232996/project/main/gCSI_GRdata_v1.3.tsv/data/human_suppl.txt"

mouse_file <- "https://raw.githubusercontent.com/s232996/project/main/gCSI_GRdata_v1.3.tsv/data/mouse.csv"

mouse_suppl_file <- "https://raw.githubusercontent.com/s232996/project/main/gCSI_GRdata_v1.3.tsv/data/mouse_suppl.txt"
```

```{r message = FALSE}
#The following functions download the files and saves them in #the data correct folders

human_data <- download_csv(human_file, "01_dat_load_human")
mouse_data <- download_csv(mouse_file, "01_dat_load_mouse")

human_suppl_data <- download_txt(human_suppl_file, "01_dat_load_human_suppl")
mouse_suppl_data <- download_txt(mouse_suppl_file, "01_dat_load_mouse_suppl")
```

```{r warning = FALSE}
#Combine downloaded data into a list for further processing


data_list <- list(human_data, human_suppl_data, mouse_data, mouse_suppl_data)

head_all(data_list)# display the head of every dataset contained in the datalist
```

\
