==> R/05_analysis_1.qmd <==
---
title: "Analysis 1 - Analysis of the mouse dataset"
editor: visual
echo : false
warning: false
format: 
  html:
    embed-resources: true
---

```{r message=FALSE, echo =FALSE, warning = FALSE}
#|label: load the packages
#install.packages("VennDiagram")
library(VennDiagram)
library(tidyverse)
library(stringr)
library(dplyr)
library(tidyr)
library(broom)
library(ggrepel)
library(RColorBrewer)
source("../R/99_proj_func.R")
```

This analysis will see which genes are significantly up and down regulated in the treatment trials

Here we begin by loading the augmented mouse dataset

```{r echo = FALSE, message = FALSE}
Sys.setenv(VROOM_CONNECTION_SIZE = 3062144) # Setting a larger connection buffer to be able to load the data
# Loading the data
mouse_df <- read_tsv(file = "../data/03_dat_aug_mouse.tsv.gz")
```

Now lets do an analysis to see which genes are significantly up and down regulated in each of the treatment trials.

```{r}
mouse_G007LK_lm <- generate_lm_genes(mouse_df, 'G007LK')
signif_genes_error_bars(mouse_G007LK_lm, 'G007LK')
```

```{r echo = FALSE}
volcano_plot(mouse_G007LK_lm, 'G007LK')
```

```{r echo = FALSE}
mouse_WNT_lm <- generate_lm_genes(mouse_df, 'WNT')
signif_genes_error_bars(mouse_WNT_lm, 'WNT')

```

```{r echo = FALSE}
volcano_plot(mouse_WNT_lm, 'WNT')
```

```{r echo = FALSE}
mouse_WNTG007LK_lm <- generate_lm_genes(mouse_df, 'WNT_G007LK')
signif_genes_error_bars(mouse_WNTG007LK_lm, 'WNT_G007LK')

```

```{r echo = FALSE}
volcano_plot(mouse_WNTG007LK_lm, 'WNT_G007LK')
```

Lets also make a Venn diagram of which significantly regulated genes are present in multiple treatments

```{r echo = FALSE}
genes_G007LK <- mouse_G007LK_lm %>% 
  filter(signif) %>%
  pull(Gene)

genes_WNTG007LK <- mouse_WNTG007LK_lm %>% 
  filter(signif) %>% 
  pull(Gene)

genes_WNT <- mouse_WNT_lm %>% 
  filter(signif) %>% 
  pull(Gene)

gene_lists <- list(G007LK = genes_G007LK, WNTG007LK = genes_WNTG007LK, WNT = genes_WNT)

```

```{r echo = FALSE}
myCol <- brewer.pal(3, "Pastel2")
venn.diagram(
  x = gene_lists,
  category.names = c("G007LK", "WNT_G007LK", "WNT"),
  filename = "../results/05_key_plot_3.png",
  output = TRUE,

    # Output features
  imagetype="png" ,
  height = 480 , 
  width = 480 , 
  resolution = 300,
  compression = "lzw",

  # Circles
  lwd = 2,
  lty = 'blank',
  fill = myCol,
  
  # Numbers
  cex = .6,
  fontface = "bold",
  fontfamily = "sans",
  
  # Set names
  cat.cex = 0.3,
  cat.fontface = "bold",
  cat.default.pos = "outer",
  cat.pos = c(-27, 27, 135),
  cat.dist = c(0.055, 0.055, 0.000),
  cat.fontfamily = "sans",
  rotation = 1
)
knitr::include_graphics("../results/05_key_plot_3.png")
log_files <- list.files(path = "../results", 
                        pattern = "05_key_plot_3\\.png\\.2023.*\\.log$", 
                        full.names = TRUE)
file.remove(log_files)


#source: https://r-graph-gallery.com/14-venn-diagramm
```

\
