==> R/06_analysis_2.qmd <==
---
title: "Analysis 2 file - PCA analysis on the human dataset"
editor: visual
echo : false
warning: false
format: 
  html:
    embed-resources: true
---

## Analysis 2 - PCA analysis on the human data set 

```{r message=FALSE, echo =FALSE, warning = FALSE}
library(tidyverse) 
library(broom) 
library(dplyr) 
library(cowplot) 
library(purrr)
library(ggplot2)
library(ggrepel)

Sys.setenv(VROOM_CONNECTION_SIZE = 3062144) # Setting a larger connection buffer to be able to load the data
human_log_expression <- read_tsv(file = "../data/03_dat_aug_human_wide_gene.tsv.gz")
```

Now we are going to perform a PCA analysis on the Cell relative expression on different conditions

```{r message = FALSE}
#|cache: TRUE 
#Here we implement the model for the PCA analysis
pca_result <- human_log_expression |> 
  select_if(is.numeric) |> #selects the columns that are numeric 
  scale() |> #scales the columns 
  prcomp()
```

This is the plot corresponding the differences in expression between the different conditions

```{r echo = FALSE}
pca_augmented <- pca_result |>
  augment(data = human_log_expression) 
# add original dataset back in 
treated_cells<- pca_augmented |> 
  filter(treatment == "G007LK")

pca_augmented |> 
  ggplot(aes(.fittedPC1, 
             .fittedPC2, 
             color = treatment,
             shape = treatment)) + 
    geom_point(size = 3) + 
  scale_color_manual(values  = c("G007LK" = "blue", "DMSO" ="red"))+
  theme_half_open(12) +
  geom_text_repel(data = treated_cells, 
                  mapping = aes(label = cell_line)) +
  theme_half_open(12) +
  background_grid() + 
  labs(x = "PC1(10%)", 
       y = "PC2(8.2%)", 
       title = "PCA Plot Depicting Cell line Responses to G007.LK") 

ggsave("../results/06_key_plot_1.png", bg = "white", device = "png")
log_files <- list.files(path = "../results", pattern = "pca_diagram\\.png\\.2023.*\\.log$", full.names = TRUE)
```

Rotation matrix

```{r eval =FALSE, echo = FALSE}
pca_result |> 
  tidy(matrix = "rotation")
```

```{r echo = FALSE, eval =FALSE}
pca_result |> 
  tidy(matrix = "eigenvalues")
```

```{r echo =FALSE}
pca_result |> 
  tidy(matrix = "eigenvalues") |>
  ggplot(aes(PC, percent)) + 
  geom_col(fill = "#56B4E9", alpha = 0.8) +
  scale_x_continuous(breaks = 1:9) + 
  scale_y_continuous(labels = scales::percent_format(), 
                     expand = expansion(mult = c(0, 0.01))) +
  theme_minimal_hgrid(12)

```

The PCA plot generated by our PC analysis depicts a stratification of the cell lines mostly by the first principal component. In addition, we can see, that for the most part, the treated cells with G007.LK are depicted to the left of the cells that belong to the same cell line and were developed in the DMSO solvent. This stratification here, though not to drastic, can give us some insights on the behavior of the cells when we give them the G007.LK . The more obvious segregation of the cells based on their line on suggests that biological characteristics have a more detrimental effect in the gene expression rather than the effect of the treatment. Though. our results can convey limited insights due to the relatively low variation that is captured by the first and the second principal component.
