==> R/07_analysis_3.qmd <==
---
title: "Analysis 3 - PCA analysis on the human dataset"
editor: visual
echo : false
warning: false
format: 
  html:
    embed-resources: true
---

## Analysis 3 - PCA analysis on the human data set (log2foldchange) 

```{r message=FALSE, echo =FALSE, warning = FALSE}
library(tidyverse) 
library(dplyr)
library(broom) 
library(ggplot2)
library(cowplot) 
library(purrr)
library(scales)
library(ggrepel)

Sys.setenv(VROOM_CONNECTION_SIZE = 3062144)
human_log2fold <- read_tsv(file = "../data/03_dat_aug_human_log2fc_wide.tsv.gz")
```

```{r}
#Apply the PCA model
pca_result<- human_log2fold |> 
  select_if(is.numeric) |> 
  scale() |> 
  prcomp()
```

This is the plot corresponding the Log2fold change of the cell lines

```{r echo = FALSE}
pca_augmented <- pca_result |>
  augment(data = human_log2fold) 
  
pca_augmented |>   
  ggplot(mapping = aes(x =.fittedPC1, 
             y = .fittedPC2, 
             color = cell_line))+
  geom_point(size =2) +
    geom_text_repel( 
                  mapping = aes(label = cell_line))+
  background_grid() + 
  labs( x = "PC1(12.7%)", 
        y = "PC2(7.5%)", 
        title = "PCA Plot Depicting Log2fold Change under G007.LK")

ggsave("../results/07_key_plot_1.png", bg = "white", device = "png")
```

```{r eval = FALSE, echo = FALSE}
#Rotation matrix
pca_result |> 
  tidy(matrix = "rotation")
```

```{r eval = FALSE, echo = FALSE}
pca_result |> 
  tidy(matrix = "eigenvalues")
```

This is a barplot corresponding to the data variation captured by the Principal components

```{r echo = FALSE}
pca_result |> 
  tidy(matrix = "eigenvalues") |> 
  ggplot(aes(PC, percent)) + 
  geom_col(fill = "#56B4E9", alpha = 0.8) + 
  scale_x_continuous(breaks = 1:9) + 
  scale_y_continuous( labels = scales::percent_format(), 
                      expand = expansion(mult = c(0, 0.01)) ) + 
  theme_minimal_hgrid(12)

```

The PC analysis suggests that the PC2 plays a more influential role in the cell stratification based on their log2fold change. The identified outliers (Skmel28, A375) and the distinct positioning of WM793 might harbor crucial information about unique gene expression profiles for responding to the G007.LK. One limitation to our analysis is the limited percentage of variation that is explained by the first and second principal components.
